base macro file
This code sets up two folders and compares all the documents within those two folders. It sets the variable fldrVersion1 and fldrVersion2 for the folders, strVersion1 and strVersion2 for the file strings, docVersion1 and docVersion2 for documents, and docCompareTarget for a compare target. It also creates a FileDialog object (fd) with a msoFileDialogFolderPicker title of "Select the folder that contains the original files" that will be used to choose which folders contain the documents. The code then creates a directory called "Compared" inside of both fldrVersion1 and fldrVersion2, then loops through each directory to open each document in them (strfldrversion1 & strversion1) and compares them with one another (Name:=docversion2), saves them into the Compared directory as an ActiveDocument, closes both documents without saving any changes, then repeats this process until all documents have been compared.